Json Editor - Styling Fix
Fixed the materialize CSS of the tree and the toolbar of the example
editor by copying the needed parts of materialize.css (mainly buttons
and nav).

Note: Cannot use the whole materialize.css because this leads to display
errors in combination with the JsonForms material renderers that do not
used materialize.css but Material UI.
When the editor and the tree are ported to React and Material UI this is
no longer needed.