[Editor] Improve Label Mapping capabilities
The values of a value mapping can now be either specified as a string or as an object.
In case of a string the behavior does not change. This guarantees old label mappings to behave as always.
The object allows to specify two properties:
constant: A constant value that is always displayed
property: A property whose content is displayed
If both are specified, the constant value is displayed followed by a space and the property value.